+++ app/Http/Controllers/Auth/ResetPasswordController.php
@@ -6,8 +6,6 @@ use Illuminate\Foundation\Auth\ResetsPasswords;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Password;
-use Thinktomorrow\Chief\Admin\Authorization\ChiefPasswordBroker;
-use Thinktomorrow\Chief\Admin\Authorization\ChiefPasswordBrokerResolver;
 use Thinktomorrow\Chief\App\Http\Controllers\Controller;
 
 class ResetPasswordController extends Controller
@@ -45,7 +43,7 @@ class ResetPasswordController extends Controller
 
     protected function broker()
     {
-        return (new ChiefPasswordBrokerResolver(app()))->resolve();
+        return Password::broker('chief');
     }
 
     public function redirectTo(): string
@@ -72,7 +70,7 @@ class ResetPasswordController extends Controller
         // If the password was successfully reset, we will redirect the user back to
         // the application's home authenticated view. If there is an error we can
         // redirect them back to where they came from with their error message.
-        return $response == ChiefPasswordBroker::PASSWORD_RESET
+        return $response == Password::PASSWORD_RESET
             ? $this->sendResetResponse($request, $response)
             : $this->sendResetFailedResponse($request, $response);
     }